+++ src/main/java/com/mapper/excel/converter/DoubleTypeConverter.java
@@ -1,41 +1,40 @@
 /*
  * code https://github.com/jittagornp/excel-object-mapping
  */
-package com.blogspot.na5cent.exom.converter;
+package com.mapper.excel.converter;
 
 import java.math.BigDecimal;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * @author redcrow
  */
-public class IntegerTypeConverter implements TypeConverter<Integer> {
+public class DoubleTypeConverter implements TypeConverter<Double> {
 
-    private static final Logger LOG = LoggerFactory.getLogger(IntegerTypeConverter.class);
+    private static final Logger LOG = LoggerFactory.getLogger(DoubleTypeConverter.class);
     
     @Override
-    public Integer convert(Object value, String... pattern) {
+    public Double convert(Object value, String... pattern) {
         if (value == null) {
             return null;
         }
 
-        if (value instanceof Integer) {
-            return (Integer) value;
+        if (value instanceof Double) {
+            return (Double) value;
         }
 
         if (value instanceof String) {
             //try {
-                return Integer.valueOf(((String) value).trim());
+                return Double.valueOf(((String) value).trim());
             //} catch (Exception ex) {
-            //    LOG.warn(null, ex);
+            //   LOG.warn(null, ex);
             //    return null;
             //}
         }
 
         if (value instanceof BigDecimal) {
-            return ((BigDecimal) value).intValue();
+            return ((BigDecimal) value).doubleValue();
         }
 
         return null;
