+++ src/main/java/com/mapper/excel/converter/IntegerTypeConverter.java
@@ -1,40 +1,41 @@
 /*
  * code https://github.com/jittagornp/excel-object-mapping
  */
-package com.blogspot.na5cent.exom.converter;
+package com.mapper.excel.converter;
 
 import java.math.BigDecimal;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * @author redcrow
  */
-public class DoubleTypeConverter implements TypeConverter<Double> {
+public class IntegerTypeConverter implements TypeConverter<Integer> {
 
-    private static final Logger LOG = LoggerFactory.getLogger(DoubleTypeConverter.class);
+    private static final Logger LOG = LoggerFactory.getLogger(IntegerTypeConverter.class);
     
     @Override
-    public Double convert(Object value, String... pattern) {
+    public Integer convert(Object value, String... pattern) {
         if (value == null) {
             return null;
         }
 
-        if (value instanceof Double) {
-            return (Double) value;
+        if (value instanceof Integer) {
+            return (Integer) value;
         }
 
         if (value instanceof String) {
             //try {
-                return Double.valueOf(((String) value).trim());
+                return Integer.valueOf(((String) value).trim());
             //} catch (Exception ex) {
-            //   LOG.warn(null, ex);
+            //    LOG.warn(null, ex);
             //    return null;
             //}
         }
 
         if (value instanceof BigDecimal) {
-            return ((BigDecimal) value).doubleValue();
+            return ((BigDecimal) value).intValue();
         }
 
         return null;
